Upload requirements.txt
Apparently the requirements.txt file is required for Heroku
deployment. This information was not on the slides given by Prof. Z,
hence why this is a separate commit. This is basically a list of
requirements for our app environment, and it can be filled out by the
following command

pip freeze > requirements.txt

assuming that there does not already exist a requirements.txt
file. Anyway, I am once again burning with rage.