Move model predictions until after image verification
I am not quite sure if this is the real root of the problem, but I
noticed that nothing was happening before the image rendering except
for running the models. So, I decided to verify and display the image
before running the models, since this work should only be happening if
the image is valid anyway.

I've additionally included a more robust way of catching corrupted or
otherwise invalid images. I am pretty sure it will accept any good
images, but there might also be some false positives that slip
through. This is (in my opinion) an acceptable risk since we shouldn't
exactly have many people trying to spam with bad images.