make paddings, indentation, invisibility consistent
Lots of things in this commit, but everything is stylistic. The first
order of business is to put all of the panel text into their own
class, which allowed us to more gracefully set the padding without
resorting to manual padding-left on every panel. Then, we got rid of
the padding for the panel itself, which allowed us to center all of
the headings correctly. We also got rid of a bunch of text
indentation, because it affected heading centering.

The italicisation on the results is now swapped: predicted category is
in regular text and the top five is italicised, since that is less
important.

The images are now centered rather than left aligned. Originally, the
whitespace was way more notable when we had a bunch of small images in
a row. There are still some issues since Sam used padding to create a
space separation between the images, so perhaps in a future commit we
could consider using a CSS grid.

In the instructions page, there is now more space between the "Step"
headings and the images!

Last but not least, the checkboxes should be invisible now. :)