move to CSS flexbox for consistent panel sizes
I was really trying to make our Options and Uploads panels the same
height no matter what the screensize was, but I realised that it
needed a parent container to achieve this. But then I realised that
only sibling containers of the same class will prevent div overlap. So
I enclosed all of the "rows" in our site into its own flex-container
class. This let us figure out good margins for each panel relative to
its parent flex-container and each other (in the case of the options
and upload panels).

Maybe something we can tackle in the future is making the rendered
image be vertically centered in the remaining whitespace of the upload
panel.