perfect image display and removal
We had two different problems: 1. When the user first uploaded an
image, there was no way for them to see it before running the models,
and 2. When the user clicked on "Reset", the image wasn't cleared from
the screen, potentially leading them to believe it was still
available. Both of these problems are fixed with a bit of JS magic.

This functionality has been tested on Chrome, Edge, and Firefox. This
should fix Issue #32, 33.